1-9-2020
This commit contains files from the sandbox, im committing them so that the braches stay on the same page(IDK if its even necessary but screw it)
With the mobile app I have the application title and the "Outage Statistics" and "Lightning Viewer" as two different pages. Haven't figured out the dropdown multiselect yet.