1-13-20 Commit
I've added a dropdown 1 and a dropdown 2, i tried adding a drop down menu to both of them but its not working. I think the issue with these drop downs is the fact that it will have to be incased in the header div and need to overlay the map, but it really doesnt want to do that, and i havent been able to figure out how to do that yet.